trex: Move package to be under checkup package
Currently the trex package is only used on the executor package. In
preparation of expanding its use to the higher level checkup package,
moving trex package to be under it.
