+++ pkg/internal/checkup/trex/types.go
